ft(GET/offices/id): user should be able to get a specific party by id
- add model and contoller files
- make GET request to /offices/:id
- write unit tests for the endpoint

[delivers #63461042]
