+++ src/test/party.test.js
@@ -6,44 +6,14 @@ import server from '../server';
 chai.should();
 chai.use(chaiHttp);
 
-describe('UI routes', () => {
-  it('should send success at /', (done) => {
-    chai.request(server)
-      .get('/')
-      .end((err, res) => {
-        res.should.have.status(200);
-        done();
-      });
-  });
-
-  it('it should throw a 404 when request is made to wrong url', (done) => {
-    chai.request(server)
-      .get('/UI')
-      .end((err, res) => {
-        res.should.have.status(404);
-        done();
-      });
-  });
-
-  it('should throw a 400 for other methods on /', (done) => {
-    chai.request(server)
-      .post('/')
-      .end((err, res) => {
-        res.should.have.status(404);
-        done();
-      });
-  });
-});
-
-
 describe('/Post create political party', () => {
   const party = {
     name: 'pdp',
-    hqAddressUrl: 'folawiyo bankole street',
+    hqAddress: 'folawiyo bankole street',
     logoUrl: 'www.testurl.com',
   };
   const partyNoName = {
-    hqAddressUrl: 'folawiyo bankole street',
+    hqAddress: 'folawiyo bankole street',
     logoUrl: 'www.testurl.com',
   };
   const partyNoAddress = {
@@ -52,32 +22,32 @@ describe('/Post create political party', () => {
   };
   const partyNoLogo = {
     name: 'pdp',
-    hqAddressUrl: 'folawiyo bankole street',
+    hqAddress: 'folawiyo bankole street',
   };
 
   const partyEmptyField = {
     name: '   ',
-    hqAddressUrl: 'folawiyo bankole street',
+    hqAddress: 'folawiyo bankole street',
     logoUrl: 'www.testurl.com',
   };
   const hqEmptyField = {
     name: 'pdp',
-    hqAddressUrl: '   ',
+    hqAddress: '   ',
     logoUrl: 'www.testurl.com',
   };
   const logoEmptyField = {
     name: 'pdp',
-    hqAddressUrl: 'folawiyo bankole street',
+    hqAddress: 'folawiyo bankole street',
     logoUrl: '   ',
   };
   const invalidLogoUrl = {
     name: 'pdp',
-    hqAddressUrl: 'folawiyo bankole street',
+    hqAddress: 'folawiyo bankole street',
     logoUrl: 'hbjniomol,',
   };
   const invalidAddress = {
     name: 'pdp',
-    hqAddressUrl: 'folawiyo vevever32es',
+    hqAddress: 'folawiyo vevever32es',
     logoUrl: 'www.testurl.com',
   };
   it('it should Create a new party with correct status code', (done) => {
@@ -97,7 +67,7 @@ describe('/Post create political party', () => {
       .send(party)
       .end((err, res) => {
         res.body.data[0].should.have.include.key('name');
-        res.body.data[0].should.have.include.key('hqAddressUrl');
+        res.body.data[0].should.have.include.key('hqAddress');
         res.body.data[0].should.have.include.key('logoUrl');
         done();
       });
@@ -185,7 +155,7 @@ describe('/Post create political party', () => {
 describe('GET /parties/:id', () => {
   const party2 = {
     name: 'pdp',
-    hqAddressUrl: 'folawiyo bankole street',
+    hqAddress: 'folawiyo bankole street',
     logoUrl: 'www.testurl.com',
   };
   before((done) => {
@@ -239,7 +209,7 @@ describe('GET /parties', () => {
 describe('/patch Update party name', () => {
   const party3 = {
     name: 'pdp',
-    hqAddressUrl: 'folawiyo bankole street',
+    hqAddress: 'folawiyo bankole street',
     logoUrl: 'www.testurl.com',
   };
 
@@ -325,7 +295,7 @@ describe('/patch Update party name', () => {
 describe('DELETE a party', () => {
   const party4 = {
     name: 'pdp',
-    hqAddressUrl: 'folawiyo bankole street',
+    hqAddress: 'folawiyo bankole street',
     logoUrl: 'www.testurl.com',
   };
   beforeEach((done) => {
@@ -373,220 +343,6 @@ describe('DELETE a party', () => {
           status: 406,
           error: 'The id parameter must be a number',
         });
-        done();
-      });
-  });
-});
-
-describe('/Post create political office', () => {
-  const office = {
-    type: 'federal',
-    officeName: 'presidency',
-    age: '50',
-  };
-  const noOfficeType = {
-    officeName: 'presidency',
-    age: '50',
-  };
-  const noOfficeName = {
-    type: 'federal',
-    age: '50',
-  };
-  const noOfficeAge = {
-    type: 'federal',
-    officeName: 'presidency',
-  };
-
-  const emptyType = {
-    type: '   ',
-    officeName: 'presidency',
-    age: '50',
-  };
-  const emptyofficeName = {
-    type: 'legislative',
-    officeName: '    ',
-    age: '50',
-  };
-  const emptyAge = {
-    type: 'state',
-    officeName: 'governor',
-    age: '     ',
-  };
-  const invalidType = {
-    type: 'ioejbermob',
-    officeName: 'presidency',
-    age: '50',
-  };
-  const invalidAge = {
-    type: '   ',
-    officeName: 'presidency',
-    age: '10',
-  };
-  const invalidAgeType = {
-    type: 'local government',
-    officeName: 'councellor',
-    age: 'xx',
-  };
-  it('it should Create a new office with correct status code', (done) => {
-    chai.request(server)
-      .post('/api/v1/offices')
-      .send(office)
-      .end((err, res) => {
-        res.should.have.status(201);
-        res.body.status.should.be.equal(201);
-        done();
-      });
-  });
-
-  it('it should Create a new office with required fields', (done) => {
-    chai.request(server)
-      .post('/api/v1/offices')
-      .send(office)
-      .end((err, res) => {
-        res.body.data[0].should.have.include.key('type');
-        res.body.data[0].should.have.include.key('officeName');
-        res.body.data[0].should.have.include.key('age');
-        done();
-      });
-  });
-
-  it('it should not Create a new office without type', (done) => {
-    chai.request(server)
-      .post('/api/v1/offices')
-      .send(noOfficeType)
-      .end((err, res) => {
-        res.should.have.status(400);
-        done();
-      });
-  });
-  it('it should not Create a new office without age', (done) => {
-    chai.request(server)
-      .post('/api/v1/offices')
-      .send(noOfficeAge)
-      .end((err, res) => {
-        res.should.have.status(400);
-        done();
-      });
-  });
-  it('it should not Create a new office without office name', (done) => {
-    chai.request(server)
-      .post('/api/v1/offices')
-      .send(noOfficeName)
-      .end((err, res) => {
-        res.should.have.status(400);
-        done();
-      });
-  });
-
-  it('it should throw a 400 when there are spaces in the party office name field', (done) => {
-    chai.request(server)
-      .post('/api/v1/offices')
-      .send(emptyofficeName)
-      .end((err, res) => {
-        res.should.have.status(400);
-        done();
-      });
-  });
-  it('it should throw a 406 when there are spaces in the office type field', (done) => {
-    chai.request(server)
-      .post('/api/v1/offices')
-      .send(emptyType)
-      .end((err, res) => {
-        res.should.have.status(406);
-        done();
-      });
-  });
-
-  it('it should throw a 406 when there are spaces in the office age field', (done) => {
-    chai.request(server)
-      .post('/api/v1/offices')
-      .send(emptyAge)
-      .end((err, res) => {
-        res.should.have.status(406);
-        done();
-      });
-  });
-
-  it('it should throw a 406 when creating a party with invalid type', (done) => {
-    chai.request(server)
-      .post('/api/v1/offices')
-      .send(invalidType)
-      .end((err, res) => {
-        res.should.have.status(406);
-        done();
-      });
-  });
-
-  it('it should throw a 400 when creating a party with wrong age value < 30', (done) => {
-    chai.request(server)
-      .post('/api/v1/offices')
-      .send(invalidAge)
-      .end((err, res) => {
-        res.should.have.status(406);
-        done();
-      });
-  });
-
-  it('it should throw a 406 when the age value isNan', (done) => {
-    chai.request(server)
-      .post('/api/v1/offices')
-      .send(invalidAgeType)
-      .end((err, res) => {
-        res.should.have.status(406);
-        done();
-      });
-  });
-});
-
-describe('GET /offices', () => {
-  it('should get all existing offices', (done) => {
-    chai.request(server)
-      .get('/api/v1/offices')
-      .end((err, res) => {
-        res.should.have.status(200);
-        done();
-      });
-  });
-});
-
-
-describe('GET /offices/:id', () => {
-  const office2 = {
-    type: 'federal',
-    officeName: 'presidency',
-    age: '50',
-  };
-  before((done) => {
-    chai.request(server)
-      .post('/api/v1/offices')
-      .send(office2)
-      .end(() => {
-        done();
-      });
-  });
-  it('should get the matching office', (done) => {
-    chai.request(server)
-      .get('/api/v1/offices/1')
-      .end((err, res) => {
-        res.should.have.status(200);
-        done();
-      });
-  });
-
-  it('should return 406 on nan ids', (done) => {
-    chai.request(server)
-      .get('/api/v1/offices/sv')
-      .end((err, res) => {
-        res.should.have.status(406);
-        done();
-      });
-  });
-
-  it('should return 404 when id is not found', (done) => {
-    chai.request(server)
-      .get('/api/v1/offices/10')
-      .end((err, res) => {
-        res.should.have.status(404);
         done();
       });
   });
